Move declare_tensor_N into knossos-pybind
